refactor: split utils.ts into focused modules (criticalMoments, matchNarrative, matchTimeline, resourceSnapshot)
utils.ts is now a barrel re-export for backward compatibility. Fixes
duplicate DEATH_LOOKFORWARD_SECONDS declaration and unused imports
(isHealerSpec, specToString, buildJsonSituationSnapshot) that were
breaking tsc. Also adds ts-jest + @types/jest for TypeScript test support.
