+++ package.json
@@ -115,6 +115,7 @@
   "devDependencies": {
     "@electron/notarize": "^3.0.0",
     "@types/estree": "^0.0.52",
+    "@types/jest": "^30.0.0",
     "@types/line-reader": "^0.0.34",
     "@types/md5": "^2.3.2",
     "@types/ps-node": "^0.1.2",
@@ -127,6 +128,7 @@
     "patch-package": "^8.0.0",
     "prettier": "^3.0.3",
     "shx": "^0.3.4",
+    "ts-jest": "^29.4.9",
     "tsdx": "^0.14.1",
     "typescript": "^4.6.3",
     "wait-on": "^7.2.0"
