+++ packages/parser/jest.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   testEnvironment: 'node',
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest', { tsconfig: 'tsconfig.json' }],
+  },
 };
